+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Add ga4 tracking smart answer results ([PR #3235](https://github.com/alphagov/govuk_publishing_components/pull/3235))
 * Add GA4 tracking to our breadcrumbs ([PR #3257](https://github.com/alphagov/govuk_publishing_components/pull/3257))
 * Make Navbar Menu Relatively Positioned Instead of Absolutely Positioned ([PR #3201](https://github.com/alphagov/govuk_publishing_components/pull/3201))
+* Fix smart answer results tracking ([PR #3261](https://github.com/alphagov/govuk_publishing_components/pull/3261))
 
 ## 34.10.1
 * Remove the brand colour for Department for Energy Security and Net Zero ([PR #3255](https://github.com/alphagov/govuk_publishing_components/pull/3255))